+++ apps/meteor/app/livechat/server/api/lib/livechat.ts
@@ -57,6 +57,7 @@ export function findGuest(token: string): Promise<ILivechatVisitor | null> {
 			visitorEmails: 1,
 			department: 1,
 			activity: 1,
+			contactId: 1,
 		},
 	});
 }
@@ -142,6 +143,7 @@ export async function settings({ businessUnit = '' }: { businessUnit?: string } 
 			hiddenSystemMessages: initSettings.Livechat_hide_system_messages,
 			livechatLogo: initSettings.Assets_livechat_widget_logo,
 			hideWatermark: initSettings.Livechat_hide_watermark || false,
+			visitorsCanCloseChat: initSettings.Omnichannel_allow_visitors_to_close_conversation,
 		},
 		theme: {
 			title: initSettings.Livechat_title,