+++ packages/core-typings/src/ILivechatContact.ts
@@ -14,12 +14,17 @@ export interface ILivechatContactConflictingField {
 
 export interface ILivechatContact extends IRocketChatRecord {
 	name: string;
-	phones: string[];
-	emails: string[];
+	phones?: string[];
+	emails?: string[];
 	contactManager?: string;
 	unknown?: boolean;
 	hasConflict?: boolean;
 	conflictingFields?: ILivechatContactConflictingField[];
 	customFields?: Record<string, string | unknown>;
 	channels?: ILivechatContactChannel[];
+	createdAt: Date;
+	lastChat?: {
+		_id: string;
+		ts: Date;
+	};
 }